+++ src/main/runtime/workspace-session-terminal-membership-authority.ts
@@ -5,8 +5,8 @@ import type {
 } from '../../shared/terminal-tab-types'
 import type { WorkspaceSessionState } from '../../shared/workspace-session-state-types'
 import { getRepoIdFromWorktreeId } from '../../shared/worktree/id'
+import { layoutContainsLeafId } from '../persistence/restoring-sessions/terminal-layout-normalization'
 import { pruneTabGroupLayoutAfterRetirement } from './mobile-session-terminal-retirement'
-import { getTerminalLeafMembershipIndex } from './terminal-leaf-membership-index'
 
 function collectLeafIds(node: TerminalPaneLayoutNode | null, ids: Set<string>): void {
   if (!node) {
@@ -165,12 +165,25 @@ export function advanceTerminalTopologyRevision(
  * The tab whose live layout holds this leaf. Only the leaf half of a pane key is remint-stable —
  * `detachTerminalPaneToTab` moves a live pane into a new tab, so a stored tabId names the tab the
  * pane left. Callers fencing on location must resolve it here rather than trust a frozen tabId.
+ *
+ * Allocation-free and stateless on purpose: writers graft leaves by assigning into a layout that is
+ * already inside the layouts record, so any cache here would need a revalidation key that is itself
+ * O(tabs) per read — the same cost as this walk, with a staleness invariant to keep.
  */
 export function findTerminalTabIdForLeaf(
   session: WorkspaceSessionState | undefined,
   leafId: string
 ): string | undefined {
-  return getTerminalLeafMembershipIndex(session?.terminalLayoutsByTabId).get(leafId)
+  const layouts = session?.terminalLayoutsByTabId
+  if (!layouts) {
+    return undefined
+  }
+  for (const tabId of Object.keys(layouts)) {
+    if (layoutContainsLeafId(layouts[tabId]?.root ?? null, leafId)) {
+      return tabId
+    }
+  }
+  return undefined
 }
 
 export function hasHostAuthoritativeTerminalMembership(
